Add files via upload
Overall the project uses the scan method to accept user input into an ArrayList. It does this by executing a while loop 10 times, checking against a condition to ensure the values entered are appropriate. It then performs simple math operations on the list obtaining a minimum, maximum, and avaerage value from the inputted numbers.